+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Changed
 
 - Remove rocket and do all communication via websockets
+- Remove "generate_config" feature flag and use curl for fetching of files
 
 ## 0.3.0 - 2025-02-16